+++ frontend/src/pages/Checkout.tsx
@@ -1,3 +1,4 @@
+import { useEffect } from "react"
 import { useTranslation } from "react-i18next"
 import { Container } from "react-bootstrap"
 import { Link } from "react-router-dom"
@@ -5,6 +6,10 @@ import { useParams } from "react-router-dom"
 const Checkout = () => {
   const { t } = useTranslation()
   const { lang } = useParams()
+
+  useEffect(() => {
+    window.scrollTo(0, 0)
+  }, [])
 
   return (
     <Container className="min-vh-100">